# default.openapi .json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Default module",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/stocks": {
      "post": {
        "summary": "create",
        "deprecated": false,
        "description": "",
        "tags": [
          "stocks"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "stock_name": "Soybean",
                "quantity": 100,
                "unit": "kg"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "message": "Stock created"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "get": {
        "summary": "get-all",
        "deprecated": false,
        "description": "",
        "tags": [
          "stocks"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": false,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "message": "Stocks fetched successfully",
                  "data": [
                    {
                      "stock_id": 1,
                      "stock_name": "Soybean",
                      "quantity": 200,
                      "unit": "kg",
                      "updated_at": "2026-03-07T10:30:00"
                    },
                    {
                      "stock_id": 2,
                      "stock_name": "Sugar",
                      "quantity": 150,
                      "unit": "kg",
                      "updated_at": "2026-03-07T10:30:00"
                    }
                  ],
                  "pagination": {
                    "page": 1,
                    "limit": 10,
                    "total": 2,
                    "total_pages": 1
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/400 Bad Request",
            "description": ""
          },
          "401": {
            "$ref": "#/components/responses/401 Unauthorized",
            "description": ""
          },
          "500": {
            "$ref": "#/components/responses/500 Internal Server Error",
            "description": ""
          }
        },
        "security": []
      }
    },
    "/stocks/{id}": {
      "get": {
        "summary": "get-one",
        "deprecated": false,
        "description": "",
        "tags": [
          "stocks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "message": "Stock fetched successfully",
                  "data": {
                    "stock_id": 1,
                    "stock_name": "Soybean",
                    "quantity": 200,
                    "unit": "kg"
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/400 Bad Request",
            "description": ""
          },
          "401": {
            "$ref": "#/components/responses/401 Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/404 Not Found",
            "description": ""
          },
          "500": {
            "$ref": "#/components/responses/500 Internal Server Error",
            "description": ""
          }
        },
        "security": []
      },
      "put": {
        "summary": "update",
        "deprecated": false,
        "description": "",
        "tags": [
          "stocks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "stock_name": "Soybean",
                "quantity": 250,
                "unit": "kg"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": {
                  "message": "Stock updated successfully",
                  "data": {
                    "stock_id": 1,
                    "stock_name": "Soybean",
                    "quantity": 250,
                    "unit": "kg"
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/400 Bad Request",
            "description": ""
          },
          "401": {
            "$ref": "#/components/responses/401 Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/404 Not Found",
            "description": ""
          },
          "500": {
            "$ref": "#/components/responses/500 Internal Server Error",
            "description": ""
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}